\name{BiocAnnotRes2015-package}
\alias{BiocAnnotRes2015-package}
\alias{BiocAnnotRes2015}
\docType{package}
\title{
\packageTitle{BiocAnnotRes2015}
}
\description{
\packageDescription{BiocAnnotRes2015}
}
\author{
\packageAuthor{BiocAnnotRes2015}

Maintainer: \packageMaintainer{BiocAnnotRes2015}
}
\keyword{ package }
\examples{
}
